Feat: integreated frontend
