+++ api/index.js
@@ -1,0 +1,94 @@
+const db = require('../db')
+const express = require('express')
+const bcrypt = require('bcrypt-nodejs')
+
+const router = express.Router()
+
+router.post('/signin', (req, res) => {
+    const { email, password } = req.body
+    if (!email || !password) {
+      return res.status(400).json('incorrect form submission')
+    }
+    db.select('email', 'hash')
+      .from('login')
+      .where('email', '=', email)
+      .then((data) => {
+        const isValid = bcrypt.compareSync(password, data[0].hash)
+        if (isValid) {
+          return db
+            .select('*')
+            .from('users')
+            .where('email', '=', email)
+            .then((user) => {
+              res.json(user[0])
+            })
+            .catch((err) => res.status(400).json('unable to get user'))
+        } else {
+          res.status(400).json('wrong credentials')
+        }
+      })
+      .catch((err) => res.status(400).json('wrong credentials'))
+  })
+  
+  router.post('/register', (req, res) => {
+    const { email, name, password } = req.body
+    if (!email || !name || !password) {
+      return res.status(400).json('incorrect form submission')
+    }
+    const hash = bcrypt.hashSync(password)
+    db.transaction((trx) => {
+      trx
+        .insert({
+          hash: hash,
+          email: email,
+        })
+        .into('login')
+        .returning('email')
+        .then((loginEmail) => {
+          return trx('users')
+            .returning('*')
+            .insert({
+              email: loginEmail[0],
+              name: name,
+              joined: new Date(),
+            })
+            .then((user) => {
+              res.json(user[0])
+            })
+        })
+        .then(trx.commit)
+        .catch(trx.rollback)
+    }).catch((err) => {
+      console.error(err)
+      res.status(400).json('unable to register')
+    })
+  })
+  
+  router.get('/profile/:id', (req, res) => {
+    const { id } = req.params
+    db.select('*')
+      .from('users')
+      .where({ id })
+      .then((user) => {
+        if (user.length) {
+          res.json(user[0])
+        } else {
+          res.status(400).json('Not found')
+        }
+      })
+      .catch((err) => res.status(400).json('error getting user'))
+  })
+  
+  router.put('/image', (req, res) => {
+    const { id } = req.body
+    db('users')
+      .where('id', '=', id)
+      .increment('entries', 1)
+      .returning('entries')
+      .then((entries) => {
+        res.json(entries[0])
+      })
+      .catch((err) => res.status(400).json('unable to get entries'))
+  })
+
+  module.exports = router
