+++ server.js
@@ -2,7 +2,6 @@ const express = require('express')
 const bodyParser = require('body-parser')
 const bcrypt = require('bcrypt-nodejs')
 const cors = require('cors')
-const knex = require('knex')
 const morgan = require('morgan')
 
 const db = require('./db')
@@ -14,98 +13,12 @@ app.use(morgan('dev'))
 app.use(cors())
 app.use(bodyParser.json())
 
-app.get('/', (req, res) => {
-  res.send({
-    message: 'Hey👋🏻',
-  })
-})
+app.use(express.static('build'))
 
-app.post('/signin', (req, res) => {
-  const { email, password } = req.body
-  if (!email || !password) {
-    return res.status(400).json('incorrect form submission')
-  }
-  console.log(db)
-  db.select('email', 'hash')
-    .from('login')
-    .where('email', '=', email)
-    .then((data) => {
-      const isValid = bcrypt.compareSync(password, data[0].hash)
-      if (isValid) {
-        return db
-          .select('*')
-          .from('users')
-          .where('email', '=', email)
-          .then((user) => {
-            res.json(user[0])
-          })
-          .catch((err) => res.status(400).json('unable to get user'))
-      } else {
-        res.status(400).json('wrong credentials')
-      }
-    })
-    .catch((err) => res.status(400).json('wrong credentials'))
-})
+app.use('/api', require('./api/index'))
 
-app.post('/register', (req, res) => {
-  const { email, name, password } = req.body
-  if (!email || !name || !password) {
-    return res.status(400).json('incorrect form submission')
-  }
-  const hash = bcrypt.hashSync(password)
-  db.transaction((trx) => {
-    trx
-      .insert({
-        hash: hash,
-        email: email,
-      })
-      .into('login')
-      .returning('email')
-      .then((loginEmail) => {
-        return trx('users')
-          .returning('*')
-          .insert({
-            email: loginEmail[0],
-            name: name,
-            joined: new Date(),
-          })
-          .then((user) => {
-            res.json(user[0])
-          })
-      })
-      .then(trx.commit)
-      .catch(trx.rollback)
-  }).catch((err) => {
-    console.error(err)
-    res.status(400).json('unable to register')
-  })
-})
-
-app.get('/profile/:id', (req, res) => {
-  const { id } = req.params
-  db.select('*')
-    .from('users')
-    .where({ id })
-    .then((user) => {
-      if (user.length) {
-        res.json(user[0])
-      } else {
-        res.status(400).json('Not found')
-      }
-    })
-    .catch((err) => res.status(400).json('error getting user'))
-})
-
-app.put('/image', (req, res) => {
-  const { id } = req.body
-  db('users')
-    .where('id', '=', id)
-    .increment('entries', 1)
-    .returning('entries')
-    .then((entries) => {
-      res.json(entries[0])
-    })
-    .catch((err) => res.status(400).json('unable to get entries'))
+app.get('/.*/', (req, res) => {
+  res.render('index.html')
 })
 
 const port = process.env.PORT || 3000
